Fix compilation of passenger for different ruby versions
With ruby 2.1.2 the directory that previously was being hardcoded into
override['rbenv']['ruby']['gems_path'] doesn't follow the same convention
we used to generate the path where gems are installed to.
So now we run a few rbenv commands on the server to find out the proper
paths for gems and other attributes needed for passenger.
It should work with any ruby version from now on.